### 2021-01-22 #2983 @whedon
[PRE REVIEW]: UltraNest - a robust, general purpose Bayesian inference engine
**Submitting author:** @JohannesBuchner (<a href="http://orcid.org/0000-0003-0426-6634">Johannes Buchner</a>)
**Repository:** <a href="https://github.com/JohannesBuchner/UltraNest/" target ="_blank">https://github.com/JohannesBuchner/UltraNest/</a>
**Version:** 3.2.0
