### 2021-01-22 #2983 @kthyng
@JohannesBuchner Your paper is too long. Please check out our [paper requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and trim yours down where possible.
