+++ paper.bib
@@ -3234,3 +3234,11 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2013PASP..125..306F},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+@misc{ultranestjs,
+  author = {Buchner, Johannes},
+  title = {ultranest-js},
+  year = {2018},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/JohannesBuchner/ultranest-js}
+}
