+++ paper.md
@@ -35,8 +35,6 @@ UltraNest is especially useful for multi-modal or non-Gaussian parameter spaces,
 computational expensive models, in robust pipelines. 
 Parallelisation to computing clusters and resuming incomplete runs is available.
 
-# Outline
-
 # Statement of need
 
 When scientific models are compared to data, two tasks are important: 
@@ -52,18 +50,16 @@ Current and upcoming large astronomical surveys require characterising
 a large number of highly diverse objects, which requires reliable analysis pipelines.
 This is what UltraNest was developed for.
 
-# Overview
-
-Nested sampling [@Skilling2004] allows Bayesian inference on arbitrary user-defined likelihoods.
+Nested sampling [NS, @Skilling2004] allows Bayesian inference on arbitrary user-defined likelihoods.
 Additional to computing parameter posterior samples, 
 it also estimates the marginal likelihood (“evidence”, $Z$).
 Bayes factors between two competing models $B=Z_1/Z_2$ are 
 a measure of the relative prediction parsimony of the models, 
 and form the basis of Bayesian model comparison.
 By performing a global scan of the parameter space from the 
-worst to best fits to the data, nested sampling performs well also in multi-modal settings.
+worst to best fits to the data, NS performs well also in multi-modal settings.
 
-In the last decade, several variants of nested sampling have been developed. 
+In the last decade, several variants of NS have been developed. 
 These include (1) how better and better fits are found while 
 respecting the priors,
 (2) whether it is allowed to go back to worse fits and explore the parameter space more,
@@ -72,10 +68,9 @@ UltraNest develops novel, state-of-the-art techniques for all of the above.
 They are especially remarkable for being free of tuning parameters and 
 theoretically justified.
 
-Currently available efficient nested sampling implementations such as `MultiNest` [@Feroz2009] and its
-open-source implementations (e.g., `nestle`, `dynesty`) rely on 
-the heuristic multi-ellipsoidal rejection algorithm which has shown biases
- when the likelihood contours are not ellipsoidal [@Buchner2014stats;@Nelson2020].
+Currently available efficient NS implementations such as `MultiNest` [@Feroz2009] and its
+open-source implementations rely on a heuristic algorithm which has shown biases
+when the likelihood contours are not ellipsoidal [@Buchner2014stats;@Nelson2020].
 UltraNest instead implements better motivated self-diagnosing algorithms,
 and improved, conservative uncertainty propagation.
 In other words, UltraNest prioritizes robustness and correctness, and maximizes 
@@ -84,297 +79,91 @@ is willing to invest computation for obtaining a
 gold-standard exploration of the entire posterior distribution in one run,
 UltraNest was developed.
 
-This package provides feature parity with other packages
- (such as `MultiNest`), e.g., circular parameters, 
- resuming incomplete runs, efficient multi-core 
-computation on clusters, and provides additional convenience features
-for visualisation and diagnostics.
-
-![The logo of UltraNest is a hedgehog cautiously walking up a likelihood function; made by https://www.flaticon.com/authors/freepik](docs/static/logo.png)
+![Logo of UltraNest; made by https://www.flaticon.com/authors/freepik](docs/static/logo.png)
 
 # Method
 
-Nested sampling methods are systematically reviewed in Buchner et al., submitted,
+NS methods are systematically reviewed in Buchner et al., submitted,
 highlighting also the approaches used in UltraNest.
 
-The basic outline of vanilla nested sampling [see @Skilling2004 for details] is as follows:
+The basic outline of vanilla NS [see @Skilling2004 for details] is as follows:
 
-A set of $N$ live points is drawn from the prior. 
-The unit hypercube is used as a natural space (u-space) and inverse cumulative prior transforms convert the point coordinates
[…]
