### 2021-01-23 7b1758c
add ref for beta sampling; word count: 994
