### 2021-02-01 45d4f6b
simplify writing csv file
