### 2021-02-01 #3001 @mattpitkin
@JohannesBuchner for the conda installation instructions in the documentation can you add that you need to explicitly specify the `conda-forge` channel, e.g., either:

```
conda install -c conda-forge ultranest
```

or

```
conda config --add channels conda-forge
conda install ultranest
```
