### 2021-02-02 #3001 @mattpitkin
@johannesbuchner can you merge the `joss-paper` branch of the ultranest repo into master so that the paper can be compiled?
