### 2021-02-02 #3001 @JohannesBuchner
I'd rather not, I understood I can use a branch https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements

> Your paper (paper.md and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default).
