### 2021-02-04 c251606
workaround for scipy inconsistency
