### 2021-02-04 #3001 @mattpitkin
@johannesbuchner can you add the DOI https://doi.org/10.1021/acs.jctc.9b00567 for the Griffiths & Wales paper? Can you also update the Higson, Handley, Hobson & Lasenby paper to the published version as given [here](https://ui.adsabs.harvard.edu/abs/2019S%26C....29..891H/abstract)?
