### 2021-02-04 #3001 @mattpitkin
Minor paper comments:
* line 16: typo "contraining" to "constraining"
* line 17: remove comma after "open source"
* line 31: either remove "also" or move it to read "also performs well"
* line 32: maybe change to say "The variants include developments related to: ..." or something similar.
* line 46: remove "also"
* line 49: explain what a "live point" is.
* line 49: "A inverse" to "An inverse"
* line 50: "converts" to "converts these live points", and "and the likelihood L evaluated." to "and the likelihood L is evaluated at each point."
* line 69: "estimate" to "estimates"
* line 71: should "integrators" be "integrator"? If not change "integrators is" to "integrators are".
