### 2021-02-06 1008f47
add conda env info for binder
