+++ environment.yml
@@ -1,0 +1,14 @@
+name: example-environment
+channels:
+  - conda-forge
+dependencies:
+  - python
+  - numpy
+  - cython
+  - pip
+  - pandas
+  - h5py
+  - corner
+  - flake8
+  - coveralls
+  - pytest-html
