+++ environment.yml
@@ -1,10 +1,12 @@
-name: example-environment
+name: binder-environment
 channels:
   - conda-forge
 dependencies:
   - python
   - numpy
   - cython
+  - scipy
+  - ultranest
   - pip
   - pandas
   - h5py
