### 2021-02-06 9edb815
simplify pump_region/stepsampler calls
