### 2021-02-07 dd3a50b
keep region_nodes better in sync
