### 2021-02-17 d39f511
Fix verbose options
"outputkwargs" dictionary should be passed when calling the "run" method, not when initializing the sampler. Otherwise calling to "pymultinest_solve_compat" fails if verbose=False
