### 2021-02-21 #3001 @ziatdinovmax
Here you go:

In their package/paper, J. Buchner introduces a general-purpose Bayesian-inference package for parameter estimation and model comparison. I was able to easily install the package and reproduce the examples in tutorials. The paper itself is written in a clear and concise manner. My only suggestion is to improve documentation of individual functions according to one of the accepted styles (reST, Google-style, Numpydoc, etc.). For example, currently, most of the package's functions just have this

```
"""This function does this and that"""
```
But it should be expanded to something like this:
```
"""
This function does this and that.

Args:
    param1: This is the first param.
    param2: This is a second param.

Returns:
    This is a description of what is returned.
"""
```
After this I will be happy to recommend the acceptance of this paper.
