### 2021-02-22 #3001 @JohannesBuchner
I haven't had time yet to address Matt's comment, but just as a quick reply: This convention is from PEP 257, section "One-line Docstrings". I use it for simple functions where I believed the function name, the argument names and the one-line docstring give sufficient documentation. If you have a list of functions which are unclear, I am happy to expand the documentation.
