### 2021-02-25 #3001 @fboehm
@JohannesBuchner - Would it be feasible to update the documentation as suggested by @ziatdinovmax? I get the impression that the suggestion is more comprehensive than what is currently written. THanks!
