### 2021-03-08 d3eaa7d
address paper comment by @mattpitkin
https://github.com/openjournals/joss-reviews/issues/3001#issuecomment-773481581
https://github.com/openjournals/joss-reviews/issues/3001#issuecomment-773492019
