+++ paper.bib
@@ -2651,25 +2651,23 @@
   url = {http://www.sciencedirect.com/science/article/pii/S1051200417301719}
 }
 @ARTICLE{Higson2017,
-  author = {{Higson}, E. and {Handley}, W. and {Hobson}, M. and {Lasenby}, A.},
-  title = {{Dynamic nested sampling: an improved algorithm for parameter estimation
-	and evidence calculation}},
-  journal = {ArXiv e-prints},
-  year = {2017},
-  month = apr,
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-  adsurl = {http://adsabs.harvard.edu/abs/2017arXiv170403459H},
-  archiveprefix = {arXiv},
-  eprint = {1704.03459},
-  file = {arXiv v1:Higson2017-eprintv1.pdf:PDF},
-  keywords = {Statistics - Computation, Astrophysics - Instrumentation and Methods
-	for Astrophysics, Physics - Data Analysis, Statistics and Probability,
-	Statistics - Methodology},
-  owner = {user},
-  primaryclass = {stat.CO},
-  
-  timestamp = {2017.04.25}
+       author = {{Higson}, Edward and {Handley}, Will and {Hobson}, Mike and {Lasenby}, Anthony},
+        title = "{Dynamic nested sampling: an improved algorithm for parameter estimation and evidence calculation}",
+      journal = {Statistics and Computing},
+     keywords = {Statistics - Computation, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Data Analysis, Statistics and Probability, Statistics - Methodology, Statistics - Computation, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Data Analysis, Statistics and Probability, Statistics - Methodology},
+         year = 2019,
+        month = sep,
+       volume = {29},
+       number = {5},
+        pages = {891-913},
+          doi = {10.1007/s11222-018-9844-0},
+archivePrefix = {arXiv},
+       eprint = {1704.03459},
+ primaryClass = {stat.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019S&C....29..891H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 @ARTICLE{Lavie2017,
   author = {Lavie, Baptiste and Mendon{\c{c}}a, Jo{\~a}o M and Mordasini, Christoph
 	and Malik, Matej and Bonnefoy, Micka{\"e}l and Demory, Brice-Olivier
@@ -2868,6 +2866,7 @@
   owner = {user},
   publisher = {ACS Publications},
   timestamp = {2021.01.07},
+  doi = {10.1021/acs.jctc.9b00567},
   url = {https://www.researchgate.net/profile/Matthew_Griffiths12/publication/336097602_Nested_Basin-Sampling/links/5dd52651a6fdcc37897ad881/Nested-Basin-Sampling.pdf}
 }
 @ARTICLE{Higson2019,
