### 2021-03-09 #3001 @mattpitkin
@JohannesBuchner thanks for the updates. If you could just address https://github.com/openjournals/joss-reviews/issues/3001#issuecomment-771025395 I'll sign-off my final tick box for the review.
