### 2021-03-09 beac1d8
add conda install channel
