+++ ultranest/stepsampler.py
@@ -21,6 +21,10 @@ def generate_random_direction(ui, region, scale=1):
         current region (not used)
     scale: float:
         length of direction vector
+    
+    Returns
+    --------
+    new direction vector
     """
     del region
     v = np.random.normal(0, 1, size=len(ui))
@@ -35,8 +39,10 @@ def generate_cube_oriented_direction(ui, region):
     -----------
     region: MLFriends object
         current region (not used)
-    scale: float:
-        length of direction vector
+
+    Returns
+    --------
+    new direction vector
     """
     del region
     ndim = len(ui)
@@ -51,8 +57,21 @@ def generate_cube_oriented_direction(ui, region):
 def generate_region_oriented_direction(ui, region, tscale=1, scale=None):
     """Draw a random direction vector in direction of one of the `region` axes.
 
-    The vector length is `scale` (if given).
+    If given, the vector length is `scale`.
     If not, the vector length in transformed space is `tscale`.
+
+    Parameters
+    -----------
+    region: MLFriends object
+        current region
+    scale: float
+        length of direction vector in u-space
+    tscale: float
+        length of direction vector in t-space
+
+    Returns
+    --------
+    new direction vector (in u-space)
     """
     ndim = len(ui)
     ti = region.transformLayer.transform(ui)
@@ -73,6 +92,17 @@ def generate_region_random_direction(ui, region, scale=1):
     """Draw a direction vector in a random direction of the region.
 
     The vector length is `scale` (in unit cube space).
+
+    Parameters
+    -----------
+    region: MLFriends object
+        current region
+    scale: float:
+        length of direction vector
+    
+    Returns
+    --------
+    new direction vector
     """
     ti_orig = region.transformLayer.transform(ui)
 
@@ -98,6 +128,9 @@ def generate_mixture_random_direction(ui, region, scale=1, uniform_weight=1e-6):
     scale: float
         length of the vector.
 
+    Returns
+    --------
+    new direction vector
     """
     v1 = generate_random_direction(ui, region)
     v1 /= (v1**2).sum()**0.5
@@ -125,7 +158,21 @@ def _inside_region(region, unew, uold):
 
 
 def inside_region(region, unew, uold):
-    """Check if `unew` is inside region."""
+    """Check if `unew` is inside region.
+
+    Parameters
+    -----------
+    region: MLFriends object
+        current region
+    unew: array
+        point to check
+    uold: array
+        not used
+    
+    Returns
+    --------
+    boolean whether point is inside the region
+    """
     del uold
     return region.inside(unew)
 
@@ -219,14 +266,20 @@ class StepSampler(object):
             self.logstat_labels += ['jump-distance', 'reference-distance']
 
     def __str__(self):
-        """Get string representation."""
         if not self.adaptive_nsteps:
             return type(self).__name__ + '(nsteps=%d)' % self.nsteps
         else:
             return type(self).__name__ + '(adaptive_nsteps=%s)' % self.adaptive_nsteps
 
     def plot(self, filename):
-        """Plot sampler statistics."""
+        """Plot sampler statistics.
+
+        Parameters
+        -----------
+        filename: str
+            Stores plot into ``filename`` and data into
+            ``filename + ".txt.gz"``.
+        """
         if len(self.logstat) == 0:
             return
 
@@ -250,11 +303,11 @@ class StepSampler(object):
         plt.close()
 
     def move(self, ui, region, ndraw=1, plot=False):
-        """Move around ui. Stub to be implemented."""
+        """Move around point ``ui``. Stub to be implemented by child classes."""
         raise NotImplementedError()
 
     def adjust_outside_region(self):
-        """Adjust proposal given that we landed outside region."""
[…]
