+++ _sources/installation.rst.txt
@@ -38,7 +38,7 @@ If you use conda, you can install UltraNest with:
 
 .. code-block:: console
 
-    $ conda install ultranest
+    $ conda install --channel conda-forge ultranest
 
 From sources
 ------------
