### 2021-03-12 #3001 @JohannesBuchner
> @JohannesBuchner thanks for the updates. If you could just address [#3001 (comment)](https://github.com/openjournals/joss-reviews/issues/3001#issuecomment-771025395) I'll sign-off my final tick box for the review.

I think this is fixed now at https://johannesbuchner.github.io/UltraNest/installation.html
