### 2021-03-12 #3001 @JohannesBuchner
> @JohannesBuchner - Would it be feasible to update the documentation as suggested by @ziatdinovmax? I get the impression that the suggestion is more comprehensive than what is currently written. THanks!

I believe I addressed this now. Please let me know of any remaining issues.

The following modules should not be considered, as they are in development and not referenced in the remaining documentation or paper: dychmc, dyhmc, flatnuts, pathsampler.
