+++ ultranest/netiter.py
@@ -41,9 +41,8 @@ class TreeNode(object):
             used to order nodes
         id: int
             refers to the order of discovery and storage (PointPile)
-        children: list of nodes
+        children: list of :class:TreeNode objects
             children nodes. if None, a empty list is used.
-
         """
         self.value = value
         self.id = id
@@ -91,9 +90,10 @@ class BreadthFirstIterator(object):
 
         Returns
         --------
-        None if done.
-        rootid, node, (active_nodes, active_root_ids, active_node_values, active_node_ids)
-        otherwise
+        tuple or None:
+            None if done.
+            `rootid, node, (active_nodes, active_root_ids, active_node_values, active_node_ids)`
+            otherwise
 
         """
         if self.active_nodes == []:
@@ -121,9 +121,15 @@ class BreadthFirstIterator(object):
         assert len(self.active_nodes) == len(self.active_node_values)
 
     def expand_children_of(self, rootid, node):
-        """Replace the current node with its children.
+        """Replace the current node with its children in the iterators
+        list of active nodes.
 
-        rootid and node have to come from the most recent call to next_node.
+        Parameters
+        ----------
+        rootid: int
+            index of the root returned by the most recent call to :pyfunc:BreadthFirstIterator.next_node
+        node: :pyclass:TreeNode
+            node returned by the most recent call to :pyfunc:BreadthFirstIterator.next_node
         """
         # print("replacing %.1f" % node.value, len(node.children))
         i = self.next_index
@@ -163,7 +169,13 @@ def _stringify_lanes(lanes, char='║'):
 
 
 def print_tree(roots, title='Tree:'):
-    """Print a pretty yet compact graphic of the tree."""
+    """Print a pretty yet compact graphic of the tree.
+
+    Parameters
+    ----------
+    roots: list of :pyclass:TreeNode
+        tree
+    """
     print()
     print(title)
     explorer = BreadthFirstIterator(roots)
@@ -206,7 +218,17 @@ def print_tree(roots, title='Tree:'):
 
 
 def dump_tree(filename, roots, pointpile):
-    """Write a copy of the tree to a HDF5 file."""
+    """Write a copy of the tree to a HDF5 file.
+
+    Parameters
+    ----------
+    filename: str
+        output filename
+    roots: list of :pyclass:TreeNode
+        tree to store
+    pointpile: :class:PointPile
+        information on the node points
+    """
     import h5py
 
     nodes_from_ids = []
@@ -235,7 +257,20 @@ def dump_tree(filename, roots, pointpile):
 
 
 def count_tree(roots):
-    """Return the maximum number of parallel edges and the total number of nodes."""
+    """Return the total number of nodes and maximum number of parallel edges.
+
+    Parameters
+    ----------
+    roots: list of :pyclass:TreeNode
+        tree
+
+    Returns
+    --------
+    count: int
+        total number of nodes
+    maxwidth: int
+        maximum number of active/parallel nodes encountered
+    """
     explorer = BreadthFirstIterator(roots)
     nnodes = 0
     maxwidth = 0
@@ -251,14 +286,24 @@ def count_tree(roots):
 
 
 def count_tree_between(roots, lo, hi):
-    """Build basic tree statistics.
+    """Return the total number of nodes and maximum number of parallel edges,
+    but only considering a interval of the tree.
+
+    Parameters
+    ----------
+    roots: list of :pyclass:TreeNode
+        tree
+    lo: float
+        lower value threshold
+    hi: float
+        upper value threshold
 
     Returns
     --------
     nnodes: int
-        the total number of nodes in the value interval lo .. hi (inclusive).
+        total number of nodes in the value interval lo .. hi (inclusive).
     maxwidth: int
-        the maximum number of parallel edges
[…]
