### 2021-03-13 #3001 @fboehm
The DOIs look good. I checked all of them manually after downloading the pdf of the paper.
