### 2021-03-25 #3001 @JohannesBuchner
OK, I think I did it:

v3.1.0

released at https://github.com/JohannesBuchner/UltraNest/releases/tag/v3.1.0

zenodo record: https://zenodo.org/record/4636924

doi: 10.5281/zenodo.4636924
