### 2021-03-28 fa0bf6a
fix derived parameters: results, plotting and affine transform
