+++ _sources/ultranest.rst.txt
@@ -28,6 +28,14 @@ ultranest.flatnuts module
    :undoc-members:
    :show-inheritance:
 
+ultranest.hotstart module
+-------------------------
+
+.. automodule:: ultranest.hotstart
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 ultranest.integrator module
 ---------------------------
 
