+++ _sources/example-warmstart.ipynb.txt
@@ -4,12 +4,13 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "# Tutorial: experimenting with models\n",
+    "# Tutorial: Warm and hot start for rapid iterations\n",
     "\n",
     "In this tutorial you will learn:\n",
     "\n",
     " - How to play with model variations\n",
-    " - How UltraNest can resume and reuse an existing run, even if you modify the data/likelihood\n",
+    " - Warm start: How UltraNest can resume and reuse an existing run, even if you modify the data/likelihood\n",
+    " - Hot start: How you can make UltraNest skip ahead to the posterior peak\n",
     "\n",
     "As a simple example, lets say we want to estimate the mean and standard deviation of a sample of points. Over time, more and more points are added."
    ]
@@ -55,8 +56,29 @@
    "source": [
     "import matplotlib.pyplot as plt\n",
     "\n",
-    "plt.figure()\n",
-    "plt.errorbar(x=np.arange(Ndata), y=y, yerr=sigma_true, marker='x', ls=' ');\n"
+    "plt.figure(figsize=(10, 5))\n",
+    "plt.errorbar(x=np.arange(Ndata), y=y, yerr=sigma_true, marker='x', ls=' ');"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "We will ingest the data in chunks, with more and more information becoming available to us. Here are the chunks. We will first analyse the orange ones:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "plt.figure(figsize=(10, 5))\n",
+    "plt.errorbar(x=np.arange(Ndata), y=y, yerr=sigma_true, marker='x', ls=' ')\n",
+    "plt.errorbar(x=np.arange(Ndata)[:10], y=y[:10], yerr=sigma_true, marker='x', ls=' ')\n",
+    "ymin, ymax = plt.ylim()\n",
+    "plt.vlines(np.arange(10, Ndata, 20), ymin, ymax, linestyles='--', color='gray')\n",
+    "plt.ylim(ymin, ymax);"
    ]
   },
   {
@@ -92,7 +114,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "## Adding one data point at a time, no warm start"
+    "## Adding one new chunk at a time, no warm start"
    ]
   },
   {
@@ -110,6 +132,13 @@
     "    sampler_ref = ReactiveNestedSampler(parameters, log_likelihood, prior_transform)\n",
     "    res_ref = sampler_ref.run(min_num_live_points=400, max_num_improvement_loops=0, viz_callback=None, frac_remain=0.5)\n",
     "    reference_results.append(res_ref)\n"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Warm start"
    ]
   },
   {
@@ -151,7 +180,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "## Likelihood evaluations saved"
+    "## Likelihood evaluations saved by warm start"
    ]
   },
   {
@@ -187,18 +216,260 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
+    "## Take-aways:\n",
+    "\n",
+    "Notice the time saving in the bottom panel by more than half. This benefit is *independent of problem dimension*. The cost savings are higher, the more similar the modified problem is."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Hot start"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
[…]
