### 2021-03-31 #3001 @fboehm
> So why is there a giant pile of missing or invalid dois?

@kthyng - I verified that the dois in the generated pdf work. Should I also be concerned with the output of whedon's check on references? I had thought that it was imperfect. I don't know the algorithm that whedon uses for reference check. Does it look a the bib file? Do we need the bib file to contain only those references that are used in paper.md? Or is it ok if the bib file has unused references?
