### 2021-03-31 #3001 @JohannesBuchner
> @JohannesBuchner I'm looking at your Zenodo archives and they look out of order. The most recent one is associated with an older version of the code. Can I verify this is the correct DOI even though Zenodo tells me there is a more recent one when I click on it: 10.5281/zenodo.4636924.

Thanks, I think zenodo got confused there because I did a bunch at the same time.

The latest version is now 3.2.0.
doi: 10.5281/zenodo.4653423
