### 2021-03-31 #3001 @JohannesBuchner
> @JohannesBuchner Just a couple of notes for your paper:
>
>     * there is a typo in the 2nd paragraph: chose > choose.
>
>     * why are there extra periods in the [2] affiliation line?

Thank you, these are fixed in the updated branch.
