### 2021-03-31 14fdc9f
fixed invalid dois which were urls
