+++ paper.bib
@@ -1174,7 +1174,7 @@
   volume = {102},
   pages = {878 - 886},
   number = {4},
-  doi = {https://doi.org/10.1016/j.bpj.2011.12.053},
+  doi = {10.1016/j.bpj.2011.12.053},
   issn = {0006-3495},
   owner = {user},
   
@@ -1305,7 +1305,7 @@
   volume = {1443},
   pages = {145-156},
   number = {1},
-  doi = {http://dx.doi.org/10.1063/1.3703630},
+  doi = {10.1063/1.3703630},
   file = {Published version:Skilling2012.pdf:PDF},
   owner = {user},
   timestamp = {2014.02.19},
@@ -2374,7 +2374,7 @@
   year = {2016},
   volume = {201},
   pages = {8 - 18},
-  doi = {http://dx.doi.org/10.1016/j.cpc.2015.12.005},
+  doi = {10.1016/j.cpc.2015.12.005},
   issn = {0010-4655},
   keywords = {Nested Sampling},
   owner = {user},
@@ -2642,7 +2642,7 @@
   year = {2017},
   volume = {70},
   pages = {84 - 93},
-  doi = {http://dx.doi.org/10.1016/j.dsp.2017.07.021},
+  doi = {10.1016/j.dsp.2017.07.021},
   file = {:henderson2017combined.pdf:PDF},
   issn = {1051-2004},
   publisher = {Elsevier},
