### 2021-04-02 #3001 @kthyng
@fboehm @JohannesBuchner As long as the DOIs in the paper are ok, it is fine. I would be remiss not to check in though with the unusual listing of doi's coming through our whedon check.
